ParametricInversion.jl

Parametric Inversion is a library for inverting Julia programs.

Quick Start

using ParametricInversion
f(x, y, z) = x + y * z
x_, y_, z_ = rand(3)
output = f(x_, y_, z_)
params = rand(2)
(x__, y__, z__) = invertinvoke(f, Tuple{Float64, Float64, Float64}, output, params)
f(x__, y__, z__)

Index

ParametricInversion.ParametricInversion

ParametricInversion.Loc

ParametricInversion.PIConstant

ParametricInversion.Places

ParametricInversion.add!

ParametricInversion.branchpreds

ParametricInversion.branchvars

ParametricInversion.choose

ParametricInversion.cycle

ParametricInversion.duplify!

ParametricInversion.enumplaces

ParametricInversion.fargs

ParametricInversion.head

ParametricInversion.incomingbranches

ParametricInversion.inverseparameters

ParametricInversion.invert

ParametricInversion.invert!

ParametricInversion.invertci

ParametricInversion.invertinvoke

ParametricInversion.invertir

ParametricInversion.knownvars

ParametricInversion.locallyundefinedvars

ParametricInversion.nused

ParametricInversion.nused

ParametricInversion.passvars!

ParametricInversion.producedvars

ParametricInversion.reversestatements!

ParametricInversion.stmtargtypes

ParametricInversion.stmts

ParametricInversion.usedvars

ParametricInversion.varreplace!

ParametricInversion.varstatements

ParametricInversion.vartypes

ParametricInversion.ParametricInversion — Module

Julia in reverse: Parametric Inversion in pure Julia

ParametricInversion.Loc — Type

Location in trace

ParametricInversion.PIConstant — Type

Constant

ParametricInversion.Places — Type

Places refers to a subset of the axes (aka dimension, column, attributes) of relation

ParametricInversion.add! — Method

add v to vm[k]

ParametricInversion.branchpreds — Method

All

ParametricInversion.branchvars — Method

Vars that are used in branching in block b

ParametricInversion.choose — Function

choose(ctx, loc, f, types, target, known. θ)

Parametrically choose a value for variables in relation defined by method f(::types..), given information about other values in relation.

Arguments

f(::types...) - method to choose from

loc - Location

target - which axis we wish to choose onto

given - what is given/known about values of the relation (e.g. value of input)

θ - Parameter values which determine which element to choose

ParametricInversion.cycle — Method

cycle(f, args...) yields xs_ such f(xs_...) == f(args...)`

ParametricInversion.duplify! — Method

Ensure there is single usage of a particular variable

ParametricInversion.enumplaces — Method

For statement of the form v1 = f(v2, v3, ..., vn) produces [v1, v2, ..., vn]

ParametricInversion.fargs — Method

Arguments of a statement

ParametricInversion.head — Method

Head of expression defined by statement

ParametricInversion.incomingbranches — Method

What branchblocks are incoming into block b

ParametricInversion.inverseparameters — Function

Parameters that can be used to invert to x from y parameters(f, types, args...)

Produces parameters θ such that:

inverseparameters(f, types, f(args...), θ) == args

These parameters can be used for learning.

ParametricInversion.invert! — Method

invert block b, store result in invb, assume v ∈ knownvars is known

ParametricInversion.invert — Method

ir::IR that computes inverse inverse of ir

ParametricInversion.invertci — Method

Produce inverse ::CodeInfo

ParametricInversion.invertinvoke — Method

invertinvoke(f, t::Type{<:Tuple}, z, θ)

Parametric inverse application of method f(::t...) to args with parameters θ

Inputs:

f - function to invert t - Tuple of types which determines method of f to invert z - Input to inverse method θ - Parameter values

Returns

(a, b, c, ...) - tuple of values

f(x, y, z) = x * y + z
x, y, z = invertinvoke(f, Tuple{Float64, Float64, Float64}, 2.3, rand(3))
@assert f(x, y, z) == 2.3

ParametricInversion.invertir — Method

invertir(f, t::Type{T})

Produce inverse ast for method f(::T)

ParametricInversion.knownvars — Method

When entering block at branchpoint brid, which variables are known?

ParametricInversion.locallyundefinedvars — Method

Variables that are used in block b but neither defined in b nor inputs to b

ParametricInversion.nused — Method

Number of times var is used in block

ParametricInversion.nused — Method

Num times v is used in smt

ParametricInversion.passvars! — Method

passvars!(ir)

Removes implicit use of variable. Updates ir such that if a block uses some variable v then v is an input to that block.

ParametricInversion.producedvars — Method

Variables that are produced by b: inputs or lhs of statements

ParametricInversion.reversestatements! — Method

Undo each operation statement %a = f(%x, %y, %z) in b, add to invb

ParametricInversion.stmtargtypes — Method

Argument types of statement stmt according to vartypes vtypes

ParametricInversion.stmts — Method

vars used in statements of b

ParametricInversion.usedvars — Method

vars used in some form by block b

ParametricInversion.varreplace! — Method

Repalce all occurances of x with y variable v1 with variable v2 in block b

ParametricInversion.varstatements — Method

Each statement of block b and associated Variable

ParametricInversion.vartypes — Method

Mapping from variables (including arguments) to inferred types